refactor(monorepo): split Python retrieval module by concern
